+++ src/main.rs
@@ -78,14 +78,20 @@ fn main() {
 
     for path in rx {
         if matches.canonical {
-            println!("{}", path.display());
+            print!("{}", path.display());
         } else if matches.relative {
             let relative = pwd.and_then(|pwd| pathdiff::diff_paths(&path, pwd));
             let p = relative.as_ref().unwrap_or(&path);
-            println!("{}", p.display());
+            print!("{}", p.display());
         } else {
             let p = path.strip_prefix(&base).unwrap_or(&path);
-            println!("{}", p.display());
+            print!("{}", p.display());
+        }
+
+        if matches.print0 {
+            print!("\0");
+        } else {
+            println!();
         }
 
         if !command.is_empty() {